[add,fix] useQueryUser, getUser 함수 파일위치 수정 및 Bookmark 에러로 수정
