+++ app/_hooks/useQueries/bookmarks/index.ts
@@ -1,8 +1,4 @@
-import {
-  getBookmark,
-  getLikes,
-  getUser,
-} from "@/app/_api/bookmark/bookmarkQueries";
+import { getBookmark, getLikes } from "@/app/_api/bookmark/bookmarkQueries";
 import { QUERY_KEY_BOOKMARK, QUERY_KEY_USER } from "@/app/_api/queryKeys";
 import { useQuery } from "@tanstack/react-query";
 
@@ -10,14 +6,6 @@ export const useBookmark = () => {
   const { data, isLoading } = useQuery({
     queryKey: [QUERY_KEY_BOOKMARK],
     queryFn: getBookmark,
-  });
-  return { data, isLoading };
-};
-
-export const useQueryUser = () => {
-  const { data, isLoading } = useQuery({
-    queryKey: [QUERY_KEY_USER],
-    queryFn: getUser,
   });
   return { data, isLoading };
 };
